Renamed AbstractDeduperOptimized to AbstractBloomFilterDeduper and cleaned up unused class variable: saveDataAtCheckpoint. Moved orderedOutput functionality to AbstractDeduper.
